Split up subcommands
