+++ perft/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+anyhow = "1.0.75"
 chess = { path = "../chess" }
 clap = { version = "4.4.7", features = ["derive"] }
 colored = "2.0.4"
